requirements.txt
xgboost==2.0.3
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
matplotlib==3.8.2
pymysql==1.1.0
ta==0.11.0
seaborn==0.13.0
scipy==1.11.4
joblib==1.3.2
tqdm==4.66.1
python-dotenv==1.0.0